+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 * passing the `$compression` argument as a string or `null` to `BackupExport::compression()` had been deprecated and has
   been removed (backwards compatibility removed);
 * the `BackupManager::index` methods no longer returns, in the array for each file, the `filename` key;
+* the `BackupExport::filename()` method now throws an `IOException` exception (rather than a `LogicException`) if the
+  target (directory) is not writable or if the filename already exists;
 * the `TestCase::createSomeBackups()` method has been improved;
 * the `AbstractBackupUtility::__get()` method is deprecated and will be removed in a future release;
 * `DeleteAllCommand`, `RotateCommand` and `SendCommand` classes had been deprecated and have been removed;